Move out utils classes
because main package becomes too convoluted
