+++ src/main/java/de/uni_potsdam/hpi/bpt/fcm2cpn/dataModel/Association.java
@@ -3,7 +3,7 @@ package de.uni_potsdam.hpi.bpt.fcm2cpn.dataModel;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import de.uni_potsdam.hpi.bpt.fcm2cpn.Pair;
+import de.uni_potsdam.hpi.bpt.fcm2cpn.utils.Pair;
 
 public class Association extends Pair<AssociationEnd, AssociationEnd> {
