+++ src/test/java/de/uni_potsdam/hpi/bpt/fcm2cpn/GeneralModelStructureTests.java
@@ -1,7 +1,7 @@
 package de.uni_potsdam.hpi.bpt.fcm2cpn;
 
-import static de.uni_potsdam.hpi.bpt.fcm2cpn.Utils.elementName;
-import static de.uni_potsdam.hpi.bpt.fcm2cpn.Utils.normalizeElementName;
+import static de.uni_potsdam.hpi.bpt.fcm2cpn.utils.Utils.elementName;
+import static de.uni_potsdam.hpi.bpt.fcm2cpn.utils.Utils.normalizeElementName;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assumptions.assumeFalse;
@@ -34,6 +34,7 @@ import org.cpntools.accesscpn.model.Place;
 import org.cpntools.accesscpn.model.Transition;
 
 import de.uni_potsdam.hpi.bpt.fcm2cpn.testUtils.ForEachBpmn;
+import de.uni_potsdam.hpi.bpt.fcm2cpn.utils.Pair;
 
 
 public class GeneralModelStructureTests extends ModelStructureTests {
